Refactor distance calculations
