+++ pirates/world.js
@@ -1,4 +1,5 @@
 import { assets } from './assets.js';
+import { iso } from './utils/distance.js';
 
 // Attempt to import simplex-noise locally for Node environments; fall back to CDN for browsers.
 let createNoise2D;
@@ -141,23 +142,6 @@ export function isoToCart(isoX, isoY, tileWidth, tileIsoHeight, tileImageHeight)
   return { x: cartX, y: cartY };
 }
 
-// Calculate the Euclidean distance between two isometric points.
-export function isoDistance(
-  aIsoX,
-  aIsoY,
-  bIsoX,
-  bIsoY,
-  tileWidth,
-  tileIsoHeight,
-  tileImageHeight
-) {
-  const a = isoToCart(aIsoX, aIsoY, tileWidth, tileIsoHeight, tileImageHeight);
-  const b = isoToCart(bIsoX, bIsoY, tileWidth, tileIsoHeight, tileImageHeight);
-  const dx = b.x - a.x;
-  const dy = b.y - a.y;
-  return Math.sqrt(dx * dx + dy * dy);
-}
-
 // Translate canvas/screen coordinates into fractional tile indices.
 export function screenToTile(
   cssX,
@@ -178,6 +162,8 @@ export function screenToTile(
     c: sy / tileIsoHeight + sx / tileWidth
   };
 }
+
+export { iso };
 
 // Backward compatibility: accept cartesian offsets and convert internally.
 export function screenToTileWithOffset(
